Lua item adding and placement working! Trying to debug certain lua lines - creating scene lua object crashes lua
