+++ OpenGLEngine/OpenGLEngine/InstanceFileLoaderAPI.cpp~RFc92da5.TMP
@@ -1,0 +1,22 @@
+#include "InstanceFileLoaderAPI.h"
+
+
+static InstanceFileLoader loader;
+
+void InstanceFileLoaderAPI::Expose(LuaContextHandle contextHandle, string luaAPIName)
+{
+	LuaContext* pContext = LuaManager::GetInstance().GetContext(contextHandle);
+	pContext->ExposeFunction(luaAPIName, "readFile", ReadFile);
+	pContext->ExposeFunction(luaAPIName, "readFromLoadedFile", ReadFromLoadedFile);
+
+}
+
+void InstanceFileLoaderAPI::ReadFile(string filePath)
+{
+	loader.LoadFile(filePath);
+}
+
+InstanceHandle InstanceFileLoaderAPI::ReadFromLoadedFile(int index)
+{
+	return loader.ReadFromLoadedFile(index);
+}
