+++ OpenGLEngine/OpenGLEngine/LuaManager.cpp
@@ -5,6 +5,8 @@
 #include "ModelLibraryAPI.h"
 #include "RenderManagerAPI.h"
 #include "MainAPI.h"
+#include "InstanceFileLoaderAPI.h"
+#include "LuaInstanceFileLoaderManager.h"
 
 LuaContext* LuaManager::GetContext(LuaContextHandle contextHandle)
 {
@@ -35,6 +37,9 @@ void LuaManager::Initialize()
   luaManager.AddAPI("modelLibraryAPI", ModelLibraryAPI::Expose);
   luaManager.AddAPI("renderManagerAPI", RenderManagerAPI::Expose);
   luaManager.AddAPI("mainAPI", MainAPI::Expose);
+  luaManager.AddAPI("instanceFileLoaderAPI", InstanceFileLoaderAPI::Expose);
+  luaManager.AddAPI("luaInstanceFileLoaderManager", LuaInstanceFileLoaderManager::Expose);
+  
 }
 
 void LuaManager::AddAPI(string apiName, LuaAPIExposeFunc exposeFunc)
