+++ OpenGLEngine/OpenGLEngine/ObjectInstanceAPI.h
@@ -10,7 +10,7 @@ class ObjectInstanceAPI
 {
 public:
 	static ObjectInstance * GetInstance(int instHandle);
-	static void SetTranslation(int instHandle, vec3 const& translation);
+	static void SetTranslation(int instHandle, float in1,float in2,float in3);
 	static void TestRender(int instHandle);
 
 	static void Expose(LuaContextHandle contextHandle, string luaAPIName);
