+++ OpenGLEngine/OpenGLEngine/OpenGLEngine.config
@@ -1,0 +1,8 @@
+<?xml version="1.0"?>
+<Project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
+  <propertyGroup>
+    <StartProgram>C:\Users\elizabeth\Documents\GitHub\Engine\OpenGLEngine\3rdParty\Lua\lua-5.3.3_Win32_bin\lua53.exe</StartProgram>
+    <StartArguments>game.lua</StartArguments>
+    <StartWorkingDirectory>C:\Users\elizabeth\Documents\GitHub\Engine\OpenGLEngine\3rdParty\Lua\lua-5.3.3_Win32_bin</StartWorkingDirectory>
+  </propertyGroup>
+</Project>
