--- OpenGLEngine/OpenGLEngine/World.lua
@@ -1,40 +1,0 @@
-world =
-{
-	local player =
-	{
-
-	}
-	
-	bodies = {}
-	
-	scenes = {}
-	
-	function world:enterScene(sceneToEnter)
-	{
-	
-	}
-	
-	function world:loadSceneFromFile(sceneFile) return 'sceneIndex'
-	{
-	newScene
-	newScene.id = scenes.count
-	}
-}
-
-scene =
-{
-local id
-
-local function scene:addRigidBody()
-{
-}
-	
-}
-
-
-RigidBody
-InteractableBody = RigidBody:New()
-NPCBody = InteractableBody:New()
-PickUpBody = InteractableBody:New()
-
-
